requirements.txt
streamlit
pillow
PyMuPDF==1.22.3
easyocr
numpy